opcua_ingestion/requirements.txt
paho-mqtt==1.6.1
psycopg2-binary==2.9.3
flask==2.1.1
requests==2.27.1
opcua==0.98.13
pymodbus==2.5.3
